modified bill input modified, bill calculated, ingredient amount updated.

+++ pkg/utils/dto/ingredientsDto.go
@@ -5,5 +5,5 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type IngredientDto struct {
 	ID       primitive.ObjectID `json:"id" bson:"_id"`
 	Name     string             `json:"ingredientName" bson:"ingredientName,omitempty"`
-	Quantity float32            `json:"quantity" bson:"quantity,omitempty"`
+	Quantity int                `json:"quantity" bson:"quantity,omitempty"`
 }
